--- src/components/common/OrderModalComponent.vue
@@ -1,75 +1,0 @@
-<template>
-    <div class="modal-overlay">
-        <div class="modal">
-        <slot></slot>
-        <!-- <div class="btn-container"> -->
-          <div>
-            <div><button class="btn" @click="cancel">닫기</button></div>
-          </div>
-        <!-- </div> -->
-        </div>
-    </div>
-</template>
-
-<script lang="ts">
-import { defineComponent } from 'vue';
-
-export default defineComponent({
-    props: {
-        visible: Boolean,
-    },
-    methods: {
-      execute() {
-        this.$emit('execute')  
-      },
-      cancel() {
-        this.$emit('cancel')
-      },
-  },
-})
-</script>
-
-<style lang="scss" scoped>
-@import '../../assets/variable.scss';
-.modal-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.modal {
-  background-color: $main--background-color;
-  padding: 20px;
-  width: 30vw;
-  // height: 50vh;
-  border-radius: 5px;
-  box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.2);
-}
-
-.btn {
-  // margin-top: 10px;
-  background-color: $main--background-color;
-  color: $main--btn-color;
-  width: 100%;
-  border: 1px solid white;
-}
-
-.btn:hover {
-  // margin-top: 10px;
-  background-color: white;
-  color: black;
-  width: 100%;
-  border: 1px solid white;
-}
-
-.btn-container {
-  margin-top: 10px;
-}
-</style>
-
