+++ src/components/store/order/OrderCart.vue
@@ -1,77 +1,95 @@
 <template>
-  <OrderOption />
-  <div class="order-cart-container">
-    <div class="order-cart-item" v-for="menu in cartItems" :key="menu.menuId">
-      <OrderDetail :menu="menu"/>
-      <OrderQuantityButton :menu="menu" />
+<div class="order-container">
+    <div class="order-container__item" v-for="menu in cartItems" :key="menu.menuId">
+        <div class="order-container__item-info">
+            <div class="order-container__item-name">
+            {{ menu.menuName }}
+            </div>
+            <div class="order-item__quantity">
+            {{ menu.quantity }}개
+            </div>
+        </div>
+        <div class="order-cart__option">
+            <div class="order-cart__btn-group">
+                <div class="order-cart__increase-btn">
+                    <button class="increase-btn btn" @click="increaseItemQuantity(menu.menuId)">
+                        +
+                    </button>
+                </div>
+                <div class="order-cart__decrease-btn btn">
+                    <button class="decrease-btn btn" @click="decreaseItemQuantity(menu.menuId)">
+                        -
+                    </button>
+                </div>
+            </div>  
+        </div>
     </div>
   </div>
-  <OrderTotalPrice :totalPrice="getTotalPrice"/>
-  <OrderButton @click="openModal"/>
-  <PaymentTypeModal :visible="isModalVisible" @noVisible="closeModal"/>
 </template>
 
 <script lang="ts">
-import { defineComponent } from "vue";
-import OrderTotalPrice from "./OrderTotalPrice.vue";
-import OrderButton from "./OrderButton.vue";
-import OrderOption from "./OrderOption.vue";
-import OrderQuantityButton from "./OrderQuantityButton.vue";
-import OrderDetail from "./OrderDetail.vue";
-import PaymentTypeModal from "./PaymentTypeModal.vue";
-import { mapGetters, mapState } from 'vuex';
-
+import { defineComponent } from 'vue';
+import { mapMutations, mapState } from 'vuex';
 
 export default defineComponent({
-  data() {
-    return {
-      cartList: [],
-      isModalVisible: false,
-    };
-  },
-  methods: {
-    closeModal() {
-      this.isModalVisible = false;
+    computed: {
+        ...mapState(['cartItems']),
     },
-    openModal() {
-      this.isModalVisible = true;
-    },
-  },
-  components: {
-    OrderOption,
-    OrderTotalPrice,
-    OrderButton,
-    OrderQuantityButton,
-    OrderDetail,
-    PaymentTypeModal
-},
-  computed: {
-    ...mapState(['cartItems']),
-    ...mapGetters(['getTotalPrice'])
-  },
-  
-});
+    methods: {
+        ...mapMutations(['increaseQuantity', 'decreaseQuantity']),
+        increaseItemQuantity(menuId: number) {
+            this.increaseQuantity(menuId)
+        },
+        decreaseItemQuantity(menuId: number) {
+            this.decreaseQuantity(menuId)
+       },
+    }
+})
 </script>
 
 <style lang="scss" scoped>
 @import '../../../assets/variable.scss';
-.order-cart-container {
+
+.order-container {
   overflow-y: scroll;
   overflow-x: hidden;
   height: 75vh;
   border-bottom: 1px solid white;
   background-color: $main--background-color;
 }
-.order-cart-item {
+.order-container__item {
   display: flex;
   color: white;
   border: 1px solid white;
   justify-content: space-between;
 }
 
-.menu-name {
-  position: relative;
-  /* left: 10px; */
+.order-container__item-info {
+    display: flex;
+    align-items: center;
+    width: 100%;
+    padding: 0 10px;
+    background-color: $main--background-color;
 }
+
+.order-container__item-name {
+    flex-grow: 9;
+    flex-basis: 0;
+}
+
+.order-item__quantity {
+    flex-grow: 2;
+    flex-basis: 0;
+    text-align: end;
+
+}
+
+.btn {
+    border: 1px solid white;
+    background: none;
+    color: white;
+    width: 5vh;
+}
+
 
 </style>
