+++ src/components/store/order/OrderTotalPrice.vue
@@ -1,34 +1,25 @@
 <template>
-    <div class="total-price">
-        총액 = {{ getTotalPrice }}
-        
+    <div class="order-total-price">
+        총액 = {{ `${totalPrice.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",")}원` }}
     </div>
 </template>
 <script lang="ts">
 import { defineComponent } from 'vue';
+import { mapState } from 'vuex';
 
 export default defineComponent({
-    props: {
-        totalPrice: Number 
-    },
     computed: {
-        getTotalPrice(){
-            
-            if(this.totalPrice != null){
-                return `${this.totalPrice.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",")}원`
-            } else {
-                return '--------'
-            }
-        }
-    }
-})
-
+        ...mapState(['totalPrice']),
+    },
+  },
+)
 
 </script>
 
 <style lang="scss" scoped>
 @import '../../../assets/variable.scss';
-.total-price {
+
+.order-total-price {
   color: white;
   border-bottom: 1px solid white;
   display: flex;
@@ -36,6 +27,5 @@ export default defineComponent({
   align-items: center;
   height: 5vh;
   background-color: $main--background-color;
-
 }
 </style>
