--- src/views/HomeView.vue
@@ -1,19 +1,0 @@
-<template>
-  <div>
-  </div>
-</template>
-
-<script lang="ts">
-import { defineComponent } from 'vue';
-
-
-export default defineComponent({
-  components: {
-
-  }
-})
-</script>
-
-<style lang="scss" scoped>
-
-</style>
